+++ src/Entity/Role.php
@@ -25,4 +25,9 @@ class Role
     {
         return $this->name;
     }
+
+    public function __toString()
+    {
+        return $this->name;
+    }
 }