+++ src/Entity/SubscriberId.php
@@ -24,4 +24,9 @@ class SubscriberId
     {
         return $this->value;
     }
+
+    public function __toString()
+    {
+        return $this->value;
+    }
 }