+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 #[2.0.3] - 2022-01-29
 ## Changed
 - `ViberApiResponseError` throws when error response returned
+### Added
+- `__toString` to value objects
 
 #[2.0.1] - 2022-01-23
 ## Changed